create ReadWriteFile incorporating methods from ReadingBook, ReadingStock, ReadingUser, WritingStock.java Delete the redundant classes
